% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{ComputeAmbientRNA}
\alias{ComputeAmbientRNA}
\title{Remove ambient RNA contamination via decontX}
\usage{
ComputeAmbientRNA(seu, split.by = NULL, cluster.name = NULL)
}
\arguments{
\item{seu}{Seurat object}

\item{split.by}{The grouping variable used to split the input object (batch correction). Default: NULL.}

\item{cluster.name}{Name of cluster field in Seurat object metadata. Must be pre-defined.}
}
\value{
Seurat object with an additional "decontX" assay (corrected counts)
  and a `decontX_contamination` column in metadata.
}
\description{
Estimate and remove ambient RNA contamination from a Seurat object
using celda::decontX. The corrected counts are stored in a new assay "decontX".
}
