% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_enrich.R
\name{.build_gsea_enrich_plots}
\alias{.build_gsea_enrich_plots}
\title{Build dotplots for RunGseaEnrich results}
\usage{
.build_gsea_enrich_plots(enricher_result, gsea_result)
}
\arguments{
\item{enricher_result}{compareCluster enricher result (or NULL).}

\item{gsea_result}{compareCluster GSEA result.}
}
\value{
A list of plot lists.
}
\description{
Build dotplots for RunGseaEnrich results
}
\keyword{internal}
