% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{.clip_log2fc_symmetric}
\alias{.clip_log2fc_symmetric}
\title{Clip log2FC to symmetric percentile range}
\usage{
.clip_log2fc_symmetric(df, fc_col = "avg_log2FC")
}
\description{
Clip log2FC to symmetric percentile range
}
\keyword{internal}
