% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{.combine_plist}
\alias{.combine_plist}
\title{Standard combine-or-return logic (internal)}
\usage{
.combine_plist(plist, combine, ncol, nrow)
}
\description{
Standard combine-or-return logic (internal)
}
\keyword{internal}
