% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{.filter_de_results}
\alias{.filter_de_results}
\title{Post-filter DE results by adjusted p-value and/or log2 fold change}
\usage{
.filter_de_results(
  markers,
  p_val_cutoff = NULL,
  logfc_cutoff = NULL,
  verbose = TRUE
)
}
\description{
Post-filter DE results by adjusted p-value and/or log2 fold change
}
\keyword{internal}
