% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{.find_knee}
\alias{.find_knee}
\title{Find the knee / elbow point using second-derivative method (internal)}
\usage{
.find_knee(var_explained)
}
\description{
Find the knee / elbow point using second-derivative method (internal)
}
\keyword{internal}
