% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{.get_top_markers_for_label}
\alias{.get_top_markers_for_label}
\title{Get top markers for labeling}
\usage{
.get_top_markers_for_label(df, cluster_levels, nlabel, features_label = NULL)
}
\description{
Get top markers for labeling
}
\keyword{internal}
