% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{.gsea_scores}
\alias{.gsea_scores}
\title{Extract GSEA running score data (internal)}
\usage{
.gsea_scores(geneList, geneSet, exponent = 1)
}
\description{
Extract GSEA running score data (internal)
}
\keyword{internal}
