% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_enrich.R
\name{.plot_regulon_drivers}
\alias{.plot_regulon_drivers}
\title{Generate plots for one pathway in FindRegulonDrivers}
\usage{
.plot_regulon_drivers(df, pw_name, label.n = 5, label.overlap = NULL)
}
\arguments{
\item{df}{Data.frame for one pathway (from result_df).}

\item{pw_name}{Pathway name.}

\item{label.n}{Top N for rank barplot.}

\item{label.overlap}{Overlap threshold for scatter labeling.}
}
\value{
List with rank_bar and scatter plots.
}
\description{
Generate plots for one pathway in FindRegulonDrivers
}
\keyword{internal}
