% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{.post_process_markers}
\alias{.post_process_markers}
\title{Post-process marker results: add gene/group/p_val_adj/test_group columns}
\usage{
.post_process_markers(
  markers,
  group1_str,
  group2_str = "others",
  p.adjust.method = "bonferroni",
  group_levels = NULL
)
}
\description{
Post-process marker results: add gene/group/p_val_adj/test_group columns
}
\keyword{internal}
