% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_module_score.R
\name{.score_aucell}
\alias{.score_aucell}
\title{AUCell scoring with batch processing to reduce memory usage}
\usage{
.score_aucell(x, gene.sets, batch.size, cores)
}
\description{
AUCell scoring with batch processing to reduce memory usage
}
\keyword{internal}
