% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{.select_top_pathways}
\alias{.select_top_pathways}
\title{Select top pathway IDs by direction and significance (internal)}
\usage{
.select_top_pathways(res_df, topTerm, direction, padjustCutoff, id_use = NULL)
}
\description{
Select top pathway IDs by direction and significance (internal)
}
\keyword{internal}
