% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{.split_composite_cols}
\alias{.split_composite_cols}
\title{Add test_group_number and test_group columns}
\usage{
.split_composite_cols(markers, group_name, split_name)
}
\description{
Add test_group_number and test_group columns
}
\keyword{+}
\keyword{Split}
\keyword{back}
\keyword{celltype}
\keyword{columns}
\keyword{composite}
\keyword{condition}
\keyword{group1/group2}
\keyword{internal}
\keyword{into}
