% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{makeContext.scmmr_raster}
\alias{makeContext.scmmr_raster}
\title{Grid makeContext method for scMMR rasterised grobs}
\usage{
\method{makeContext}{scmmr_raster}(x)
}
\arguments{
\item{x}{A grob with class \code{"scmmr_raster"}.}
}
\value{
A \code{rasterGrob}.
}
\description{
S3 method called by the grid rendering system. Renders the original grob
to an off-screen bitmap device and returns a \code{rasterGrob}.
}
