% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ribo.genes}
\alias{ribo.genes}
\title{Ribosomal gene symbols (human)}
\format{
A character vector of length 184.
}
\source{
\url{https://github.com/JarningGau/ProjectSVR}
}
\usage{
ribo.genes
}
\description{
A character vector of 184 human ribosomal protein gene symbols,
including cytoplasmic (RPL/RPS) and mitochondrial (MRPL/MRPS) ribosomal genes.
Originally from the ProjectSVR package.
}
\keyword{datasets}
